feat(new-trace): Adding issues link to trace header.
