+++ static/app/views/performance/newTraceDetails/traceDrawer/details/styles.tsx
@@ -5,16 +5,24 @@ import * as qs from 'query-string';
 import {Button as CommonButton, LinkButton} from 'sentry/components/button';
 import {DataSection} from 'sentry/components/events/styles';
 import type {LazyRenderProps} from 'sentry/components/lazyRender';
+import Link from 'sentry/components/links/link';
+import {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
 import {Tooltip} from 'sentry/components/tooltip';
 import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {Organization} from 'sentry/types/organization';
 import {trackAnalytics} from 'sentry/utils/analytics';
 import {getDuration} from 'sentry/utils/formatters';
+import {decodeScalar} from 'sentry/utils/queryString';
 import type {ColorOrAlias} from 'sentry/utils/theme';
+import useOrganization from 'sentry/utils/useOrganization';
+import {useParams} from 'sentry/utils/useParams';
 import {
   isAutogroupedNode,
+  isMissingInstrumentationNode,
   isSpanNode,
+  isTraceErrorNode,
+  isTransactionNode,
 } from 'sentry/views/performance/newTraceDetails/guards';
 import type {
   TraceTree,
@@ -285,6 +293,67 @@ function TableRow({
   );
 }
 
+function getSearchParamFromNode(node: TraceTreeNode<TraceTree.NodeValue>) {
+  if (isTransactionNode(node) || isTraceErrorNode(node)) {
+    return `id:${node.value.event_id}`;
+  }
+
+  // Issues associated to a span or autogrouped node are not queryable, so we query by
+  // the parent transaction's id
+  const parentTransaction = node.parent_transaction;
+  if ((isSpanNode(node) || isAutogroupedNode(node)) && parentTransaction) {
+    return `id:${parentTransaction.value.event_id}`;
+  }
+
+  if (isMissingInstrumentationNode(node)) {
+    throw new Error('Missing instrumentation nodes do not have associated issues');
+  }
+
+  return '';
+}
+
+function IssuesLink({
+  node,
+  children,
+}: {
+  children: React.ReactNode;
+  node?: TraceTreeNode<TraceTree.NodeValue>;
+}) {
+  const organization = useOrganization();
+  const params = useParams<{traceSlug?: string}>();
+  const traceSlug = params.traceSlug?.trim() ?? '';
+
+  const dateSelection = useMemo(() => {
+    const normalizedParams = normalizeDateTimeParams(qs.parse(window.location.search), {
+      allowAbsolutePageDatetime: true,
+    });
+    const start = decodeScalar(normalizedParams.start);
+    const end = decodeScalar(normalizedParams.end);
+    const statsPeriod = decodeScalar(normalizedParams.statsPeriod);
+
+    return {start, end, statsPeriod};
+  }, []);
+
+  return (
+    <Link
+      to={{
+        pathname: `/organizations/${organization.slug}/issues/`,
+        query: {
+          query: `trace:${traceSlug} ${node ? getSearchParamFromNode(node) : ''}`,
+          start: dateSelection.start,
+          end: dateSelection.end,
+          statsPeriod: dateSelection.statsPeriod,
+          // If we don't pass the project param, the issues page will filter by the last selected project.
+          // Traces can have multiple projects, so we query issues by all projects and rely on our search query to filter the results.
+          project: -1,
+        },
+      }}
+    >
+      {children}
+    </Link>
+  );
+}
+
 const LAZY_RENDER_PROPS: Partial<LazyRenderProps> = {
   observerOptions: {rootMargin: '50px'},
 };
@@ -343,6 +412,7 @@ const TraceDrawerComponents = {
   Duration,
   TableRow,
   LAZY_RENDER_PROPS,
+  IssuesLink,
 };
 
 export {TraceDrawerComponents};
